Move popover types to separate file
